requirements.txt
pandas
numpy
ipykernel
selenium
beautifulsoup4
requests
google-cloud-storage
spacy